+++ src/070-optimizer/interpreter.ts
@@ -57,7 +57,7 @@ import {
     isSelfId,
 } from "../050-grammar/ast";
 import { SrcInfo, dummySrcInfo, Parser, getParser } from "../050-grammar";
-import { divFloor, modFloor } from "../070-optimizer/util";
+import { divFloor, modFloor } from "./util";
 import {
     getStaticConstant,
     getStaticFunction,
